+++ deepdoc_client_action/CHANGELOG.md
@@ -3,3 +3,7 @@
 
 ## 0.0.2
 - Added file serving of uploaded documents and document management which links to vector store content. Deleting a processed document will also remove all of its references from the vector store.
+
+## 0.0.3
+- Updated action to update descriptor export on doc_manifest update
+- Updated action to allow custom metadata to be added to each document ingested